Add files via upload
The Hospital management system comprises of mainly three main modules. They are Admin 
Module, Doctor Module, and User Module. The Admin module will be responsible for adding the 
doctor details, viewing patient details, to viewing the Total appointments. The Doctor module will 
be responsible for editing the doctor's details and the total appointments to the doctor. The user 
module will be responsible for the creation of a new account and then making an appointment with 
the doctor. 
The Technologies required for this project is mainly divided into two categories. The 
frontend technologies are HTML, CSS, and bootstrap CSS, The Backend Technologies are Java, 
Servlets, JSP and For Connecting JDBC will be required. The Tools required to develop this project 
are NetBeans IDE, MYSQL workbench, and Glassfish Server.